Reorganize the code
