+++ .eslintrc.json
@@ -1,4 +1,5 @@
 {
+  "parserOptions": { "ecmaVersion": 2018 },
   "env": { "es6": true, "node": true },
   "extends": ["eslint:recommended", "prettier"],
   "plugins": ["prettier"]
