+++ src/shared/batch.js
