+++ src/shared/collect.js
