+++ src/shared/concat.js
