+++ src/shared/drain.js
