+++ src/shared/merge.js
