+++ src/shared/reduce.js
