+++ src/shared/scan.js
