+++ src/shared/take.js
